% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_standardise_indirect.R
\name{age_standardise_indirect}
\alias{age_standardise_indirect}
\title{Calculate indirectly age-standardised rates}
\usage{
age_standardise_indirect(
  interest_counts,
  interest_age_specific_population,
  standard_age_specific_counts,
  standard_age_specific_population,
  count_variable,
  population_variable,
  age_variable,
  grouping_variable,
  single_standard_population
)
}
\arguments{
\item{interest_counts}{Count of the event in the population of interest}

\item{interest_age_specific_population}{Population size for each age group in the population of interest}

\item{standard_age_specific_counts}{Count of event the for each age group in the standard population}

\item{standard_age_specific_population}{Population size for each age group in the standard population}

\item{count_variable}{Variable providing counts of the event. This must be the same in each dataset provided.}

\item{population_variable}{Variable providing population sizes. This must be the same in each dataset provided.}

\item{age_variable}{Variable providing age groups. This must be the same in each dataset provided.}

\item{grouping_variable}{Optional variable providing groups within the population. If provided, rates will be calculated separately for each group. This is useful if you wish to look at how rates vary across a variable of interest in an age-standardised way (eg: look at the rate of the event in each ethnic group while standardising to the general populaiton's age structure).}

\item{single_standard_population}{Logical value indicating whether to standardise each population group (as provided by \code{grouping_variable}) should be standardised to a single standard population. If \code{TRUE}, the groups will be standardised to the same standard population. If \code{FALSE}, the groups will be standardised to group-specific populations and the separate event counts and age group sizes must be provided for each group in the dataframes provided to \code{standard_age_specific_counts} and \code{standard_age_specific_population}.}
}
\value{
A tibble containing observed counts, expected counts and indirectly age-standardised rates of the event for the population of interest.
}
\description{
Calculates indirectly age-standardised rates of an event for a population of interest, based on the age-specific rates in a standard population.
}
\details{
This type of age standardisation requires the following four datasets:
\itemize{
\item counts of the event for the population of interest
\item population sizes in each age group for the population of interest
\item age-specific counts of the event for the standard population
\item population sizes in each age group for the standard population
}
}
\examples{
# Add Example
}
