% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_age_groups.R
\name{collapse_age_groups}
\alias{collapse_age_groups}
\title{Collapse detailed age groups into broader age groups}
\usage{
collapse_age_groups(start_groups, target_groups)
}
\arguments{
\item{start_groups}{A character vector of the detailed age groups you want to convert. These must be nested within the groups provided in \code{target_groups}.}

\item{target_groups}{A character vector of the broader age groups to which you want to convert}
}
\value{
A character vector the same length as \code{start_groups} containing the broader age groups to which each detailed age group correspond
}
\description{
Recodes a character vector of age groups into a broader set of age groups. It requires the more detailed groups to be nested within the broader groups. This can also be used to turn single years of age into age groups.
}
