% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_population_change.R
\name{decompose_population_change}
\alias{decompose_population_change}
\title{Decompose the change in a count of interest between two time points}
\usage{
decompose_population_change(
  age_specific_interest_counts,
  age_specific_population,
  count_variable,
  age_variable,
  population_variable,
  time_variable,
  start
)
}
\arguments{
\item{age_specific_interest_counts}{A dataframe of age-specific counts of the event of interest at each time point/}

\item{age_specific_population}{A dataframe of the population size in each age group at each time point.}

\item{count_variable}{A variable providing the counts in \code{age_specific_interest_counts}.}

\item{age_variable}{A variable providing the age groups in \code{age_specific_interest_counts} and \code{age_specific_population}.}

\item{population_variable}{A variable providing the population sizes in \code{age_specific_population}.}

\item{time_variable}{A variable providing the time points in \code{age_specific_interest_counts} and \code{age_specific_population}. This variable must only have two distinct values in each dataframe.}

\item{start}{The value of \code{time_variable} which indicates the first time point.}
}
\value{
A dataframe containing the raw counts attributable to each component and the observed change. An additional column indicates this raw change as a proportion (not a percent) of the observed change.
}
\description{
For an observed change between two time points in a count of interest in the population (eg: count of deaths), calculates the population growth component, the population ageing component and the rate change component.
\itemize{
\item \strong{Population growth component}: the change in the count that we would expect to see from the overall change in the size of the population.
\item \strong{Population ageing component}: the change in the count that we would expect to see from the change in the age structure of the population.
\item \strong{Rate change component}: the change in the count that is attributable to a change in the underlying likelihood of the event occurring, independent of population growth and ageing.
}
}
\examples{
# Add example
}
