% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_nssec.R
\name{ons_nssec}
\alias{ons_nssec}
\title{Convert NS-SEC variable to correctly ordered ONS categories}
\usage{
ons_nssec(
  x,
  exact = FALSE,
  detailed = TRUE,
  keep_original = FALSE,
  match_distance = 0.3
)
}
\arguments{
\item{x}{Character vector of NS-SEC values to be converted.}

\item{exact}{Logical value indicating whether to use exact matching to the ONS categories. Default is \code{FALSE}.}

\item{detailed}{Logical value indicating whether to use detailed ONS categories (8 class version). If \code{FALSE} the broad ONS categories (5 class version) will be used. Default is \code{TRUE}.}

\item{keep_original}{Logical value indicating whether to return the original NS-SEC categories provided. If \code{TRUE} the function will simply order the provided categories by the order of the matched ONS categories. Default is \code{FALSE}.}

\item{match_distance}{Maximum allowed distance for fuzzy matching. Default is \code{0.3}.}
}
\value{
A factor of NS-SEC values ordered by the order of the ONS categories.
}
\description{
Convert NS-SEC values to the closest ONS NS-SEC category and return them as an appropriately ordered factor. The ONS categories used are the 8 class or 5 class versions given in Section 7 of the \href{https://www.ons.gov.uk/methodology/classificationsandstandards/otherclassifications/thenationalstatisticssocioeconomicclassificationnssecrebasedonsoc2010#classes-and-collapses}{ONS methodology page}. If \code{exact} is \code{FALSE} (the default) fuzzy matching will be used and the user will be asked to confirm matches.
}
\examples{
# Add example
}
