% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_ageing_component.R
\name{population_ageing_component}
\alias{population_ageing_component}
\title{Calculate expected change in a count from population ageing}
\usage{
population_ageing_component(
  age_specific_interest_counts,
  age_specific_population,
  count_variable,
  age_variable,
  population_variable,
  time_variable,
  start
)
}
\arguments{
\item{age_specific_interest_counts}{A dataframe of age-specific counts of the event of interest at each time point/}

\item{age_specific_population}{A dataframe of the population size in each age group at each time point.}

\item{count_variable}{A variable providing the counts in \code{age_specific_interest_counts}.}

\item{age_variable}{A variable providing the age groups in \code{age_specific_interest_counts} and \code{age_specific_population}.}

\item{population_variable}{A variable providing the population sizes in \code{age_specific_population}.}

\item{time_variable}{A variable providing the time points in \code{age_specific_interest_counts} and \code{age_specific_population}. This variable must only have two distinct values in each dataframe.}

\item{start}{The value of \code{time_variable} which indicates the first time point.}
}
\value{
A numeric value indicating expected change
}
\description{
This is designed for internal use by the \code{\link[=decompose_population_change]{decompose_population_change()}} function. Calculates the change we expect to see in a count (eg: count of deaths) between two time points based on the ageing of the population over that period.
}
\examples{
# Add example
}
\keyword{internal}
