% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_pyramid.R
\name{population_pyramid}
\alias{population_pyramid}
\title{Plot a population pyramid}
\usage{
population_pyramid(
  data,
  count,
  age,
  side,
  left,
  proportions = FALSE,
  reference_indicator = NULL,
  reference_population = NULL,
  pal = NULL,
  colour = "black",
  linewidth = 1,
  width = 0.5,
  reference_point_size = 3,
  reference_point_linewidth = 1,
  centre_zero = TRUE
)
}
\arguments{
\item{data}{A dataframe containing counts for each age group within two subpopulations. If a reference population is also being plotted, there must also be a variable distinguishing between the main population of interest and the reference population. The dataframe must be in tidy format (ie: one row per age group-subpopulation combination, with additional rows for the reference population if being used).}

\item{count}{A variable providing population counts}

\item{age}{A variable providing age groups}

\item{side}{A variable that will define whether counts are plotted on the left or the right of the pyramid}

\item{left}{A character string which is a value of the variable provided to \code{side}. This category of \code{side} will be plotted on the left of the pyramid.}

\item{proportions}{A logical value indicating whether proportions of the total population should be plotted instead of the raw counts. If \code{TRUE} the proportion within the total population (including both subpopulations) will be plotted. If a reference population is also being plotted, the proportion is calculated within the main population and the reference population separately. Default is \code{FALSE}.}

\item{reference_indicator}{An optional variable indicating whether counts are for the main population or the reference population (only used if \code{reference_population} is not \code{NULL}). Default is \code{NULL}.}

\item{reference_population}{If not \code{NULL}, the age-structure for a reference population will be plotted on top of the pyramid bars. This should be a character string which is a value of the variable provided to \code{reference_indicator} which gives the name of the reference population. Default is \code{NULL}.}

\item{pal}{A colour palette used to distinguish the subpopulations.}

\item{colour}{A colour for the pyramid bar and reference point outlines. Default is \code{"black"}.}

\item{linewidth}{A line width for the pyramid bars outline. Default is \code{1}.}

\item{width}{A width for the pyramid bars. Default is \code{0.5}}

\item{reference_point_size}{A size for the reference points. Default is \code{3}.}

\item{reference_point_linewidth}{A line width for the reference points outline. Default is \code{1}.}

\item{centre_zero}{A logical value indicating whether to centre the x-axis around zero. Default is \code{TRUE}.}
}
\value{
A population pyramid as a \code{ggplot} object
}
\description{
Plot a population pyramid
}
\examples{
df<- data.frame(age_group = rep(c("Aged 0 to 19", "Aged 20 to 39", "Aged 40 to 59", "Aged 60 or higher"), 4),
                sex = rep(c("A", "A", "A", "A", "B", "B", "B", "B"), 2),
                n = c(rep(c(900,1100,1400,1000),2), rep(c(600,1200,1300,800),2)),
                population_type = c(rep("Population of interest", 8), rep("Reference population", 8)))

population_pyramid(data = df,
                   count = n,
                   age = age_group,
                   side = sex,
                   left = "A",
                   reference_indicator = population_type,
                   reference_population = "Reference population")

}
